+++ src/main/java/com/ex/namperfume/service/SizeService.java
@@ -3,6 +3,8 @@ package com.ex.namperfume.service;
 import com.ex.namperfume.dto.request.SizeRequest;
 import com.ex.namperfume.dto.response.SizeResponse;
 import com.ex.namperfume.entity.Size;
+import com.ex.namperfume.exception.AppException;
+import com.ex.namperfume.exception.EnumCode;
 import com.ex.namperfume.mapper.SizeMapper;
 import com.ex.namperfume.repository.SizeRepository;
 import lombok.AccessLevel;
@@ -26,15 +28,15 @@ public class SizeService {
         Size size = sizeMapper.toSize(request);
         try {
             size = sizeRepository.save(size);
-        } catch (RuntimeException e) {
-            throw new RuntimeException(e);
+        } catch (AppException e) {
+            throw new AppException(EnumCode.UNCATEGORIZE_EXCEPTION);
         }
         return sizeMapper.toSizeResponse(size);
     }
 
     public SizeResponse getSize (UUID size_id){
         return sizeMapper.toSizeResponse(sizeRepository.findById(size_id).orElseThrow(
-                ()-> new RuntimeException("Size not found in db")
+                ()-> new AppException(EnumCode.SIZE_NOT_EXIST)
         ));
     }
 
@@ -43,11 +45,13 @@ public class SizeService {
     }
 
     public void deleteSize(UUID size_id){
+        if(!sizeRepository.existsById(size_id))
+            throw new AppException(EnumCode.SIZE_NOT_EXIST);
         sizeRepository.deleteById(size_id);
     }
 
     public SizeResponse updateSizeValue(UUID size_id, int newSizeValue){
-        Size size = sizeRepository.findById(size_id).orElseThrow();
+        Size size = sizeRepository.findById(size_id).orElseThrow(()->new AppException(EnumCode.SIZE_NOT_EXIST));
         size.setSize_value(newSizeValue);
         Size updateSize = sizeRepository.save(size);
         return sizeMapper.toSizeResponse(updateSize);
